Started over in implementing baking, using my new found common sense.
