+++ src/main/java/net/pinaz993/simpleshelves/client/ShelfModel.java
@@ -1,0 +1,6 @@
+package net.pinaz993.simpleshelves.client;
+
+import net.minecraft.client.render.model.UnbakedModel;
+
+public class ShelfModel implements UnbakedModel {
+}
