+++ src/main/java/net/pinaz993/simpleshelves/woodshelves/WarpedShelf.java
@@ -1,12 +1,14 @@
-package net.pinaz993.simpleshelves;
+package net.pinaz993.simpleshelves.woodshelves;
 
 import net.minecraft.block.BlockState;
 import net.minecraft.block.entity.BlockEntity;
 import net.minecraft.util.math.BlockPos;
+import net.pinaz993.simpleshelves.AbstractShelf;
+import net.pinaz993.simpleshelves.ShelfEntity;
 import org.jetbrains.annotations.Nullable;
 
-public class BirchShelf extends AbstractShelf {
-    public BirchShelf(Settings settings) {
+public class WarpedShelf extends AbstractShelf {
+    public WarpedShelf(Settings settings) {
         super(settings);
     }
